feat(core): add a location special property
In order to have a better build experience, adding the location for importing it in the initializer
properties.

BREAKING CHANGE: The build will need to be adapted to match this new feature. The autoloader service
signature changed. Also the utils now provide a location decorator to allow you to create modules
declaring their own location.

fix #136